requirements.txt
tkinter
mysql-connector-python
pandas